feat : Critical Attribute and Modify Damage Logic

+++ Source/R1/R1Define.h
@@ -108,7 +108,7 @@ struct FR1DamageInfo
 	GENERATED_BODY()
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	float DamageAmount = 0.0f;
+	int32 DamageAmount = 0;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	FVector TargetLocation = FVector::ZeroVector;
